BUILD: fix warnings to pass werror

+++ src/txmempool.cpp
@@ -422,7 +422,7 @@ bool CTxMemPool::addUnchecked(const uint256& hash, const CTxMemPoolEntry &entry,
 
     nTransactionsUpdated++;
     totalTxSize += entry.GetTxSize();
-    if (minerPolicyEstimator && (IsArgSet("-testnet") || (entry.GetHeight()>Checkpoints::LastCheckPointHeight()))) {minerPolicyEstimator->processTransaction(entry, validFeeEstimate);}
+    if (minerPolicyEstimator && (IsArgSet("-testnet") || (entry.GetHeight() > (unsigned int)Checkpoints::LastCheckPointHeight()))) {minerPolicyEstimator->processTransaction(entry, validFeeEstimate);}
 
     vTxHashes.emplace_back(tx.GetWitnessHash(), newit);
     newit->vTxHashesIdx = vTxHashes.size() - 1;
@@ -452,7 +452,7 @@ void CTxMemPool::removeUnchecked(txiter it, MemPoolRemovalReason reason)
     mapLinks.erase(it);
     mapTx.erase(it);
     nTransactionsUpdated++;
-    if (minerPolicyEstimator && (IsArgSet("-testnet") || (it->GetHeight()>Checkpoints::LastCheckPointHeight()))) {minerPolicyEstimator->removeTx(hash, false);}
+    if (minerPolicyEstimator && (IsArgSet("-testnet") || (it->GetHeight() > (unsigned int)Checkpoints::LastCheckPointHeight()))) {minerPolicyEstimator->removeTx(hash, false);}
 }
 
 // Calculates descendants of entry that are not already in setDescendants, and adds to
@@ -595,7 +595,7 @@ void CTxMemPool::removeForBlock(const std::vector<CTransactionRef>& vtx, unsigne
             entries.push_back(&*i);
     }
     // Before the txs in the new block have been removed from the mempool, update policy estimates
-    if (minerPolicyEstimator && (IsArgSet("-testnet") || (nBlockHeight>Checkpoints::LastCheckPointHeight()))) {minerPolicyEstimator->processBlock(nBlockHeight, entries);}
+    if (minerPolicyEstimator && (IsArgSet("-testnet") || (nBlockHeight > (unsigned int)Checkpoints::LastCheckPointHeight()))) {minerPolicyEstimator->processBlock(nBlockHeight, entries);}
     for (const auto& tx : vtx)
     {
         txiter it = mapTx.find(tx->GetHash());
